+++ src/gamspy/_symbols/symbol.py
@@ -1,15 +1,8 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 import gams.transfer as gt
 
-import gamspy as gp
 import gamspy._symbols.implicits as implicits
-from gamspy.exceptions import ValidationError
-
-if TYPE_CHECKING:
-    from gamspy import Set, Alias, Parameter, Variable, Equation
 
 
 class Symbol:
@@ -18,20 +11,6 @@ class Symbol:
 
     def getStatement(self):
         """Declaration string of the symbol in GAMS"""
-
-    def _container_check(
-        self: Set | Parameter | Variable | Equation,
-        domain: list[str | Set | Alias],
-    ):
-        for set in domain:
-            if (
-                isinstance(set, (gp.Set, gp.Alias))
-                and set.container != self.container
-            ):
-                raise ValidationError(
-                    f"`Domain `{set.name}` must be in the same container"
-                    f" with `{self.name}`"
-                )
 
     def _get_domain_str(self):
         set_strs = []