+++ CHANGELOG.md
@@ -7,6 +7,7 @@ GAMSPy 0.11.8
 - General
   - Allow assigning VariableType enum or EquationType enum as an attribute after object creation for Equation and Variable.
   - Fix running twice on symbol declaration with records
+  - Return better error messages for incorrectly provided solver, options, and output arguments.
 - Testing
   - Add tests for assigning type to Variable and Equation after creation.
   - Add models information at the top of each model's docstring.